# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "AdminSys SaaS API",
    "version": "1.0.0",
    "description": "Standard SaaS Onboarding API for external integrations.",
    "contact": {
      "name": "API Support",
      "email": "support@adminsys.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Local Development Server"
    },
    {
      "url": "https://api.adminsys.com/api/v1",
      "description": "Production Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-app-key",
        "description": "Unique identifier for the SaaS application (tenant)."
      },
      "Timestamp": {
        "type": "apiKey",
        "in": "header",
        "name": "x-timestamp",
        "description": "Current Unix timestamp in seconds. Used for signature verification."
      },
      "Nonce": {
        "type": "apiKey",
        "in": "header",
        "name": "x-nonce",
        "description": "A random string unique to this request to prevent replay attacks."
      },
      "Signature": {
        "type": "apiKey",
        "in": "header",
        "name": "x-signature",
        "description": "HMAC-SHA256 hash of (Method + Path + Timestamp + Nonce + Body). Signed with App Secret."
      }
    },
...
  },
  "security": [
    {
      "ApiKeyAuth": [],
      "Timestamp": [],
      "Nonce": [],
      "Signature": []
    }
  ],
  "paths": {
...
    "/api/admin/integrations": {
      "get": {
        "tags": [
          "Admin - Integrations"
        ],
        "summary": "Get all integration configurations",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of integration configs"
          }
        }
      },
      "post": {
        "tags": [
          "Admin - Integrations"
        ],
        "summary": "Save or update an integration configuration",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Saved configuration"
          }
        }
      },
      "delete": {
        "tags": [
          "Admin - Integrations"
        ],
        "summary": "Delete an integration configuration",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/admin/product-categories": {
      "get": {
        "tags": [
          "Admin - Product Categories"
        ],
        "summary": "List product categories",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of categories"
          }
        }
      },
      "post": {
        "tags": [
          "Admin - Product Categories"
        ],
        "summary": "Create a category",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Created category"
          }
        }
      }
    },
    "/api/admin/product-categories/{id}": {
      "put": {
        "tags": [
          "Admin - Product Categories"
        ],
        "summary": "Update a category",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated category"
          }
        }
      },
      "delete": {
        "tags": [
          "Admin - Product Categories"
        ],
        "summary": "Delete a category",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/admin/system-logs": {
      "get": {
        "tags": [
          "Admin - System Logs"
        ],
        "summary": "List system logs",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of logs"
          }
        }
      }
    },
...
    "/api/ai/debates/{id}": {
      "get": {
        "tags": [
          "AI Debates"
        ],
        "summary": "Get debate details",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Debate details"
          }
        }
      },
      "delete": {
        "tags": [
          "AI Debates"
        ],
        "summary": "Delete a debate",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Debate deleted"
          }
        }
      }
    },
...
    "/api/database/export": {
      "get": {
        "tags": [
          "Database"
        ],
        "summary": "Export database SQL",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "mode",
            "schema": {
              "type": "string",
              "enum": [
                "schema",
                "bootstrap"
              ]
            }
          },
          {
            "in": "query",
            "name": "includeData",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "includeRls",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "seedTables",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "SQL dump"
          }
        }
      }
    },
...
  },
...
}